chore: migrate example app to TypeScript
When upgrading Flow (from upgrading react-native), a bunch of errors
come from the test app. Taking this opportunity to migrate to TypeScript
to avoid future issues.

Next up would be to convert the rest of the Flow files.
