+++ example/index.ts
@@ -3,9 +3,6 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @format
- * @flow
  */
 
 import {AppRegistry, Platform} from 'react-native';
