+++ tsconfig.json
@@ -1,0 +1,34 @@
+{
+  "compilerOptions": {
+    "noFallthroughCasesInSwitch": true,
+    "noImplicitOverride": true,
+    "noImplicitReturns": true,
+    "noPropertyAccessFromIndexSignature": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "strict": true,
+
+    "moduleResolution": "Node",
+    "resolveJsonModule": true,
+
+    "noEmit": true,
+
+    "allowJs": true,
+    "checkJs": true,
+
+    "allowSyntheticDefaultImports": true,
+
+    "forceConsistentCasingInFileNames": true,
+
+    "jsx": "react",
+
+    "target": "ESNext",
+
+    "skipLibCheck": true
+  },
+  "include": [
+    "example/*.ts",
+    "example/*.tsx",
+    "example/examples/**/*.tsx"
+  ]
+}
